Added SiteXML to load site xml file. Does not parse xml into a new object, instead leaves it in XML object, which is easy to traverse using E4X, making parsing unnecessary.
Changed Main.as to use SiteXML.as to get xml and create buttons. NavButton class now has an id property that gets a section title, which in turn becomes the text of the button.

Took loadSite() out of Preloader since it is never available early enough to show site loading. Could be probably be remedied by changing swf publishing settings, but will instead create a loadsite.fla/swf that will simple be a loader for main.swf.