Overhauled makingbiodiesel infographic so it is less nestedy. It was going to be too complicated to try to programatically control so many nested animations, so I undid a lot of that. It has been rewritten to have animation on main timeline, and the programming uses frame labels for the classes to use to navigate on the timeline.
I added XML parsing and made text boxes work (get text that was in XML).

Also added NavButton class so now have some working nav buttons. Need to actually create the rest of the buttons, but the class communicates with the InfographicMain to trigger loading of different sections. Need to have logic in InfographicMain that checks to see if NavButton's label id is greater or less than current label and play or skip to the appropriate frame label... This may work better with section labels that are integers, rather than decimals such as 1.1, etc.

May need to add a getter/setter in Infographic main for setting urls (so we don't run into problems loading this into another swf). This would mean that our main swf would need to build paths for it's loaded assets based on paths in the site xml file.