Added a new texture to sit on top of main.swf background but behind any content. This replaces have each swf have its own background. Also edited index.html to try improve things visually.
Added a PreloaderEvent so we can dispatch custom events from Preloader to Main.as. redispatching the loader events would not work (event.target changed when it was cloned), so this class was necessary.

Added a cool RegExp to get the file path for a loaded swf, chop off the file name and extension and return the path to the swf. This way, if the swf has a absoluteURL setter method, we can set an absolute url dynamically. Need to test this online, but works on local machine file paths.